controle_biometrico/requirements.txt
# Flask e extensões
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-APScheduler
Flask-WTF

# Comunicação com ESPHome e Home Assistant
aioesphomeapi
py-ha-ws-client

# Utilitário para variáveis de ambiente
python-dotenv

requests
bcrypt
wtforms-sqlalchemy
pytz
waitress